# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  handler.UserLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  model.AlbumStats:
    properties:
      total:
        type: integer
      totalPhotos:
        type: integer
    type: object
  model.DashboardStats:
    properties:
      albumStats:
        $ref: '#/definitions/model.AlbumStats'
      momentStats:
        $ref: '#/definitions/model.MomentStats'
      placeStats:
        $ref: '#/definitions/model.PlaceStats'
      wishStats:
        $ref: '#/definitions/model.WishStats'
    type: object
  model.MomentStats:
    properties:
      total:
        type: integer
    type: object
  model.PlaceStats:
    properties:
      total:
        type: integer
    type: object
  model.WishStats:
    properties:
      pending:
        type: integer
      total:
        type: integer
    type: object
...
  service.FileResponse:
    properties:
      id:
        type: integer
      mime_type:
        type: string
      name:
        type: string
      size:
        type: integer
      thumbnail:
        type: string
      url:
        type: string
    type: object
...
  service.FrontendSystemInfo:
    properties:
      couple:
        properties:
          boy:
            properties:
              avatar:
                $ref: '#/definitions/service.FileResponse'
              name:
                type: string
            type: object
          girl:
            properties:
              avatar:
                $ref: '#/definitions/service.FileResponse'
              name:
                type: string
            type: object
        type: object
      site:
        properties:
          description:
            type: string
          name:
            type: string
          startDate:
            type: string
        type: object
    type: object
  service.InitSystemRequest:
    properties:
      siteDescription:
        type: string
      siteName:
        type: string
      sitePassword:
        minLength: 6
        type: string
      sitePasswordConfirm:
        type: string
      startDate:
        type: string
      userAEmail:
        type: string
      userAName:
        type: string
      userAPhone:
        type: string
      userARole:
        enum:
        - boy
        - girl
        type: string
      userBEmail:
        type: string
      userBName:
        type: string
      userBPhone:
        type: string
      userBRole:
        enum:
        - boy
        - girl
        type: string
    required:
    - siteName
    - sitePassword
    - sitePasswordConfirm
    - startDate
    - userAName
    - userARole
    - userBName
    - userBRole
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /system/dashboard/stats:
    get:
      description: 获取仪表盘的统计数据，用于展示系统的整体运营情况和数据概览
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.DashboardStats'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - OAuth2Password: []
      summary: 获取仪表盘统计数据
      tags:
      - system
  /system/info:
    get:
      description: 获取系统配置的基本信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.FrontendSystemInfo'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取系统信息
      tags:
      - system
  /system/init:
    get:
      description: 检查系统是否已完成初始化
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  additionalProperties:
                    type: boolean
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 检查系统初始化状态
      tags:
      - system
    post:
      consumes:
      - application/json
      description: 初始化系统，包括创建站点信息、用户信息和设置访问密码
      parameters:
      - description: 系统初始化参数
        in: body
        name: initSystem
        required: true
        schema:
          $ref: '#/definitions/service.InitSystemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  additionalProperties:
                    type: boolean
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 初始化系统
      tags:
      - system
  /system/settings/site:
    get:
      description: 获取当前站点的设置信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - OAuth2Password: []
      summary: 获取站点设置
      tags:
      - system
    post:
      consumes:
      - application/json
      description: 更新站点的设置信息
      parameters:
      - description: 站点设置
        in: body
        name: settings
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - OAuth2Password: []
      summary: 保存站点设置
      tags:
      - system
  /user:
    get:
      consumes:
      - application/json
      description: Get user info with token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - OAuth2Password: []
      summary: Get user info
      tags:
      - user
  /user/token:
    post:
      consumes:
      - application/json
      description: User login with username and password to get token
      parameters:
      - description: User login credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Generate User Token
      tags:
      - user
  /users:
    get:
      consumes:
      - application/json
      description: Get all users list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - OAuth2Password: []
      summary: Get users list
      tags:
      - user
  /users/{id}:
    put:
      consumes:
      - application/json
      description: Update user info
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User info
        in: body
        name: user
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - OAuth2Password: []
      summary: Update user info
      tags:
      - user
  /users/{id}/avatars:
    get:
      consumes:
      - application/json
      description: Get user avatar history with pagination
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - OAuth2Password: []
      summary: Get user avatar history
      tags:
      - user
securityDefinitions:
  OAuth2Password:
    flow: password
    scopes:
      admin: Grants read and write access to administrative information
      read: Grants read access
      write: Grants write access
    tokenUrl: /api/v1/user/token
    type: oauth2
swagger: "2.0"